split LruCache and descendant classes in different compilation units
